hawk_install.sh - Make sure to use certain files only if they exists. 	Restart cron during hawk install only hawk-unblock.sh was not found in it and we just added it 	Execute /usr/local/cpanel/etc/init/stopcphulkd only if exists and only then we should try to remove /var/cpanel/cphulk_enable 	Removed commented obsolete code 	Make sure to exit cleanly at the end
uninstall_hawk.sh - Code indentation
	Removed commented code

+++ bin/hawk_install.sh
@@ -4,7 +4,7 @@
 # copyright@1h.com                                              http://1h.com
 # This code is subject to the 1H license. Unauthorized copying is prohibited.
 
-VERSION='0.1.1'
+VERSION='0.1.2'
 
 # Various paths
 syspath='/home/1h'
@@ -116,7 +116,7 @@ if ( ! cat $db | su - postgres -c "psql -Upostgres $dbname" ); then
     exit 1
 fi
 
-if ( ! grep hawk-unblock.sh /var/spool/cron/root ); then
+if [ -f /var/spool/cron/root ] && ( ! grep hawk-unblock.sh /var/spool/cron/root ); then
     if ( ! chattr -ia /var/spool/cron/root ); then
         echo "[!] chattr -ia /var/spool/cron/root FAILED"
         exit 1
@@ -125,17 +125,15 @@ if ( ! grep hawk-unblock.sh /var/spool/cron/root ); then
         echo "[!] Failed to add hawk-unblock.sh to the root cron"
         exit 1
     fi
+	if [ -d /usr/local/1h/lib/guardian/svcstop ]; then
+		touch /usr/local/1h/lib/guardian/svcstop/crond
+	fi
+	if ( ! /etc/init.d/crond restart ); then
+		echo "/etc/init.d/crond restart failed"
+		exit 1
+	fi
+	rm -f /usr/local/1h/lib/guardian/svcstop/crond
 fi
-
-if [ -d /usr/local/1h/lib/guardian/svcstop ]; then
-	touch /usr/local/1h/lib/guardian/svcstop/crond
-fi
-
-if ( ! /etc/init.d/crond restart ); then
-	echo "/etc/init.d/crond restart failed"
-	exit 1
-fi
-rm -f /usr/local/1h/lib/guardian/svcstop/crond
 
 if [ ! -f /var/lib/pgsql/data/pg_hba.conf ]; then
     echo "/var/lib/pgsql/data/pg_hba.conf is missing"
@@ -154,19 +152,14 @@ if ( ! cat /var/lib/pgsql/data/pg_hba.conf | grep -v ^$ | grep -v '\#' | awk '{p
     fi
 fi
 
-/usr/local/cpanel/etc/init/stopcphulkd
-
-if ( ! rm -rf /var/cpanel/cphulk_enable ); then
-	echo "[!] rm -rf /var/cpanel/cphulk_enable FAILED"
-	exit 1
+if [ -x /usr/local/cpanel/etc/init/stopcphulkd ]; then
+	/usr/local/cpanel/etc/init/stopcphulkd
+	if ( ! rm -rf /var/cpanel/cphulk_enable ); then
+		echo "[!] rm -rf /var/cpanel/cphulk_enable FAILED"
+		exit 1
+	fi
 fi
 
-#if [ -f $etc_dir/portal.conf ]; then
-#	. $etc_dir/portal.conf
-#	if ( ! sed -i "/window.location/s/'http.*\/\/.*'/'https:\/\/$portalmaster_url\/portal\/'/" /home/$system_user/api/cgi-bin/hawk/js/hawk.js ); then
-#		echo "Failed to set the correct link to the master interface but this is not fatal"
-#	fi
-#fi
 if ( ! /usr/local/1h/bin/lockit.sh hawk ); then
 	echo "[!] Failed to password protect the web folder with /usr/local/1h/bin/lockit.sh hawk"
 	exit 1
@@ -193,3 +186,4 @@ fi
 
 rm -rf /usr/local/1h/lib/guardian/svcstop/hawk
 
+exit 0
